backend/requirements.txt
Django==2.2.11
psycopg2-binary==2.8.4
djangorestframework==3.11.0
django-isbn-field==0.5.2